refactor(tui): split dialogs.go into co-located per-dialog files
Break the 2637-line dialogs.go into dialog_*.go feature files, each owning
its dialog's Update handler, render method, open/close, and state type:
confirm, instance, add_fleet, edit_fleet, port_forward, codespaces, session,
plus a dialog_shared.go for the text-input/nav-key helpers. Browser and
armada-selector dialogs fold into the existing browser.go / armada.go.

The dialog-overlay switch in viewFleetList becomes viewActiveDialog, a thin
dispatcher to per-dialog renderXxxDialog methods. Pure code motion +
mechanical import adjustment — no behavior change.
